AOC Day 1
Advent of code day 1. This was fun!

I'm using Claude 3.5 to teach me how to do various things in OCaml, so
treating this as an assisted run. I'm not pasting any of the problem
itself into an LLM, just using it as an easier documentation look up.
This was a simple first day, mainly just getting used to some new parts
of ocaml. I'll probably want to build some tooling that will make it
nicer to iterate on common stuff. More parsing utils.

Also to figure out is how to split these day by day. I'm not sure what
the standard dune project formatting is like.